+++ Models/ItemComboApp.cs
@@ -9,7 +9,7 @@ namespace WebAppServer.Models
     {
         public Int64 id { get; set; }
         public Int64 id_produto { get; set; }
-        public Int32 int_qtd_item { get; set; }
+        public Double int_qtd_item { get; set; }
         public Double dbl_val_unit { get; set; }
         public Int64 id_server { get; set; }
         public Int16 int_sinc { get; set; }
